Move arithmetic operators into their own module
